Refactor messageFeed to be called conversationList also set up a basic socket.io connection between messageModal and the server
