+++ client/containers/messageModal.container.js
@@ -1,6 +1,12 @@
 import messageModal from '../components/messageFeed/messageModal'
 import { connect } from 'react-redux'
 import  { toggleMessageModal, submitMessage } from '../actions/messageFeed.actions'
+import io from 'socket.io-client'
+let socket = io();
+
+socket.on('connect', function() {
+  console.log('hello world');
+})
 
 const mapStateToProps = (state, props) => {
   return {
@@ -14,6 +20,7 @@ const mapStateToProps = (state, props) => {
     person_username: state.messageFeed.person_username,
     textModalField: state.form.messageModal,
     hideConversationsInMessageModal: state.view.hideConversationsInMessageModal
+
   }
 }
 
@@ -36,6 +43,14 @@ const mapDispatchToProps = (dispatch) => {
       dispatch({
         type: 'TOGGLE_CONVERSATIONS'
       })
+    },
+    createRoom: (user1, user2) => {
+      // use the concatenation of two usernames to create a room with both users
+      const roomName = user1 > user2 ? user1 + user2 : user2 + user1;
+      // socket.emit('createRoom', roomName);
+
+
+
     }
   }
 }
