+++ package.json
@@ -32,6 +32,7 @@
     "es6-promise": "^3.1.2",
     "express": "^4.13.4",
     "history": "^2.0.1",
+    "http": "0.0.0",
     "isomorphic-fetch": "^2.2.1",
     "jimp": "^0.2.21",
     "jwt-simple": "^0.5.0",
@@ -60,6 +61,7 @@
     "redux-storage": "^4.0.0",
     "redux-storage-engine-localstorage": "^1.0.1",
     "redux-thunk": "^2.0.1",
+    "socket.io": "^1.4.5",
     "superagent": "^1.8.1",
     "webpack-dev-middleware": "^1.6.0",
     "webpack-hot-middleware": "^2.10.0"
